Renamed 'RewriteOriginal' project to 'TestSubject'
