+++ Lokad.ILPack.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lokad.ILPack", "src\Lokad.I
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lokad.ILPack.Tests", "test\Lokad.ILPack.Tests\Lokad.ILPack.Tests.csproj", "{8647DF26-57C1-465C-8E21-E44617D8EA60}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RewriteOriginal", "test\RewriteOriginal\RewriteOriginal.csproj", "{2B12E3D8-796A-4300-AE55-E151D0A129B8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestSubject", "test\TestSubject\TestSubject.csproj", "{2B12E3D8-796A-4300-AE55-E151D0A129B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
