+++ test/Lokad.ILPack.Tests/Lokad.ILPack.Tests.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Lokad.ILPack.csproj" />
-    <ProjectReference Include="..\RewriteOriginal\RewriteOriginal.csproj" />
+    <ProjectReference Include="..\TestSubject\TestSubject.csproj" />
   </ItemGroup>
 
   <ItemGroup>
