+++ test/Lokad.ILPack.Tests/dump.bat
@@ -1,4 +1,4 @@
-ildasm /bytes bin\Debug\netcoreapp2.1\RewriteOriginal.dll > il.original.txt
-ildasm /bytes bin\Debug\netcoreapp2.1\cloned\RewriteClone.dll > il.clone.txt
-mddumper bin\Debug\netcoreapp2.1\RewriteOriginal.dll > md.original.txt
-mddumper bin\Debug\netcoreapp2.1\cloned\RewriteClone.dll > md.clone.txt
+ildasm /bytes bin\Debug\netcoreapp2.1\TestSubject.dll > il.original.txt
+ildasm /bytes bin\Debug\netcoreapp2.1\cloned\ClonedTestSubject.dll > il.clone.txt
+mddumper bin\Debug\netcoreapp2.1\TestSubject.dll > md.original.txt
+mddumper bin\Debug\netcoreapp2.1\cloned\ClonedTestSubject.dll > md.clone.txt
