+++ test/TestSubject/MyClass.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 
 // This project defines a set of types that will be rewritten by the test cases to a new
-// dll named "RewriteClone".  The test cases then compare the final type information of both
+// dll named "ClonedTestSubject".  The test cases then compare the final type information of both
 // assemblies to confirm everything was re-written correction
 
 // SEE RewriteTest in Lokad.ILPack.Tests
 
 
-namespace RewriteOriginal
+namespace TestSubject
 {
     public class MyClass
     {
